Rename the files of generating uniform meshes for better understanding
